refactor: migrate getNewStatus and getStartAndEndTime to DIRACCommon
Move getNewStatus and getStartAndEndTime functions from DIRAC JobStatusUtility
to DIRACCommon as stateless utility functions. The original DIRAC module now
imports these functions from DIRACCommon.

The migrated functions intentionally retain references to undefined dependencies
(TimeUtilities, JobStatus) as part of the stateless utility migration strategy.
